PRM not yet working, testing RRT out

+++ 16350_sp20_HW2_v0/grad/utilheader.h
@@ -149,7 +149,7 @@ NodeRRT* nearestNeighbour( std::vector<double> currSamplePt_, NodeRRT* root_ );
 void treeDFS(  NodeRRT* nodeIn, std::vector<double> currSamplePt_, 
   std::priority_queue< NodePQ, std::vector<NodePQ>, CompareNN > &min_queue );
 
-int extend( NodeRRT* root_, NodeRRT* tail_, std::vector<double> currSamplePt_ , int eps_, double* map, 
+int extend( NodeRRT* root_, NodeRRT* tail_, std::vector<double> currSamplePt_ , double eps_, double* map, 
   int x_size, int y_size, std::vector<double> endCoord_ );
 
 int newConfig( std::vector<double> currSamplePt_, NodeRRT* nearestNode_, NodeRRT* newNode_ , double eps_,
